Run the tls destructors in the correct order
